requirements-test.txt
# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.20.0
pytest-xdist>=3.0.0
coverage>=6.0.0

# Testing utilities
responses>=0.22.0
httpx>=0.23.0

# Linting and code quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.990

# Type stubs
types-requests>=2.28.0
types-python-dateutil>=2.8.0
typing-extensions>=4.4.0

# Documentation
docformatter>=1.4.0

# Test data generation
factory-boy>=3.2.0
Faker>=15.0.0

# API testing
fastapi-testclient>=0.0.7
requests-mock>=1.10.0

# Database testing
pytest-postgresql>=4.1.0
pytest-sqlalchemy>=1.5.0
